Compile separate traces for each reduce() caller
Proof-of-concept solution to issue #32. Compiles the root trace for
the foldl-loop separately for each calling function. This reduces
excessive side tracing. For the specific test case from #32 it
resolves the problem perfectly and makes sure that each of the 100
traces has a nine-instruction loop.

This solution is not perfect though: if a single function makes
multiple calls to foldl then they will be compiled together. Could be
nicer to find a way to compile each call-site separately. Just have to
compute a better key to the foldl_impl table. For now this should be
sufficient for evaluating whether the optimization is beneficial in
practice.

This change uses one very recent LuaJIT feature, the 'proto' field of
debug information, so you need at least
luajit/luajit@fb61f7cbe3ec983dfc9087bde04496aa4bbaa31b.