Fix chain() to work correctly with maps
Problem: When the 2nd+ argument is a map, then `chain()` iterates only
over the first iterator.

Reason: `chain_gen_r2()` returns the next iterator only if
`2 > #param / 3`. This test does not work correctly when any of the
chain's arguments is a map - in that case the `params` table is sparse.
Why? The problem is in `gen_x, param_x, state_x = iter(elem)` -
`state_x` is always nil when `elem` is a map.

This patch fixes this problem.