doc: renew Sphinx configuration
As I see from https://luafun.github.io, the documentation was generated
last time in 2017 using Sphinx 1.4.8.

I made attempt to regenerate the documentation using the recent Sphinx
version (4.3.0) and observed the following regression.

The documentation has code blocks with highlighted lines
(`:emphasize-lines:`). With the new Sphinx version, a background color
for all lines becomes the same as background color for highlighted ones
(and so the highlight becomes invisible). The cause is [this][1] change,
but if we'll apply [this][2] recipe, the problem disappears.

[1]: https://github.com/sphinx-doc/sphinx/pull/7718
[2]: https://github.com/sphinx-doc/sphinx/pull/5282

I found the relevant update of the default sphinx-quickstart
configuration [3] and decided to renew our configuration as precaution
against all problems of this kind.

[3]: https://github.com/sphinx-doc/sphinx/pull/5700

My steps are the following:

```sh
$ cd doc
$ rm conf.py index.rst Makefile make.bat
$ sphinx-quickstart            \
    --no-sep                   \
    --project 'Lua Functional' \
    --author 'Roman Tsisyk'    \
    -v 0.1                     \
    --release 0.1.3            \
    -l en                      \
    --makefile                 \
    --no-batchfile             \
    --use-make-mode
$ git checkout -- index.rst
```

Most of values are deduced from existing configuration, but I changed
several ones:

* `-l en`: otherwise `sphinx-quickstart` asks about this and, since the
  documentation is in English, it looks appropriate.
* `--no-batchfile`: can't imagine why someone will use Windows to build
  this documentation, so I see no reason to store the batch file in the
  repository.
* `--use-make-mode`: this way `Makefile` will just pass execution to
  `sphinx-build`; it looks as modern way to build Sphinx docs and
  reduces amount of boilerplate code in the repository.

After this I made the following manual modifications:

* Restored copyright year (was 2013-2017, generated as 2021, replaced
  manually with 2013-2021).
* Restored HTML theme from 'alabaster' to 'haiku'.
* Restored `html_logo = 'logo.png'` option.
* Replaced `.. module:: fun` with `.. currentmodule:: fun` over the
  documentation sources to eliminate bunch of warnings:

  > WARNING: duplicate object description of fun, other instance in
  > basic, use :noindex: for one of them

  This way Sphinx will not assume that the module is declared several
  times. Instead, all functions will be added to the module, but the
  alphabetical index will not refer to all those pages as descriptions
  of the 'fun' module.

  Added `.. module::fun` to `reference.rst` to link here from the
  alphabetical index.
* Added `html_copy_source = False` option, because the `*.rst` sources
  are not deployed to the luafun.github.io repository (and it seems it
  would be useless).

Proceed with the following steps to build and open the documentation:

```sh
$ cd doc
$ make html
$ xdg-open _build/html/index.html
```

After all those steps the generated documentation looks mostly the same
as before. I'll note the following differences:

* The background color of code blocks becomes white (was grey).
* A font in function declarations now bigger (at least in my Web
  browser). I guess, it is due to [4].

[4]: https://github.com/sphinx-doc/sphinx/commit/9b3b8a49fb0c2c0059577ccd8993fb6ae64deb97

I'm fine with those changes.

There are some spacing and font differences, the alphabetical index is
now with bullets -- all those changes looks quite minor.

I'll update https://luafun.github.io once this patch will land.