fix drop_while work with stateful iterators
Currently, method `drop_while` rewinds the iterator once it yields
values not satisfying the condition. Such approach doesn't work with
stateful iterators since they cannot be rewound by simply copying the
state. The commit brings a new implementation that works correctly with
all types of iterators.

Performance comparison. In the first scenario, functor from `drop_while`
returns `true` on the first million elements, then it returns `false` on
the other million elements. With LuaJIT enabled, the new implementation
is about 7% faster. Without LuaJIT, results are identical. In the second
scenario, we leave only the first half of elements so that functor always
returns `true`, which results in an empty iterator since all the elements
are dropped. In this case, the new implementation is about 10-15% faster
with JIT enabled. In the third scenario, we leave the second half of
elements so that the functor returns `false` right from the start.
In this case, the new implementation is about 4% slower than the old one
with JIT enabled.

Part of #83