test preserve the state of the original iterator
Filtering and slicing functions actually simply drop some elements from
the iterator, so I would expect them to preserve its state, and I would
rely on it. For example, when using `ipairs`, I know that the first
yielded value, the state, is the index of the current element in the table,
so I could use it to determine the positions of elements after filtration
in the original table.

The commit covers all filtering and slicing functions with tests
checking if states of original iterators are preserved. Function
`take_n` doesn't work this way - see #85. Functions that are actually
composition of smaller ones (e.g, `drop` is composition of `drop_n`
and `drop_while`) are not tested.