chain: don't stop on param = nil
The implementation of the `chain` function saves the gen-param-state
triplets for all the source iterators into a table and tracks an index
in this table to know, which source iterator is the current one.

When the current iterator is exhausted, `chain`'s gen function checks
the source iterators table: is there a next source iterator?

The problem is that it checks `param` against `nil`, however it would be
more natural and correct to check `gen` against `nil`.

This commit fixes the condition, so now `chain` works for iterators that
don't use `param` at all.

Fixes #86